Support parsing structs and fields
Interfaces will be able to have fields which
are essentially class-variables.
These are just stored as values in the vtable.

Need to figure out the situation with delegates

I think somehow we need to mark DuckObject->shape as an 'implementation'
of the PyShape interface. The wrapper methods should implicitly retreive
the shape vtable from the object.

Do we also need to add some sort of 'self' type? That would seem
somewhat cleaner - but would seem like an issue with vtables.