[ast] Start to combine IvanType with TypeRef
Resolvng types will be done by modifying the original
AST in-place.
Avoids declaring the AST types for a second time.....
TODO: Only use the magic `resolved` property with `UnresolvedNamedType`
Other TypeRefs don't nessicarrily need it

Do a little bit of cleanup with the 'compiler'. The 'CodeCompiler'
will compile function bodies to C11 source code. For now we only need
this for default functions.